requirements.txt
python-dotenv~=1.0.0
py-cord~=2.4.1
aiohttp~=3.8.6
requests~=2.31.0